--- src/app/Scenes/Onboarding/useAuthenticationFormContext.tsx
@@ -1,6 +1,0 @@
-import { AuthenticationFormValues } from "app/Scenes/Onboarding/Components/AuthenticationWizard"
-import { useFormikContext } from "formik"
-
-export const useAuthenticationFormContext = () => {
-  return useFormikContext<AuthenticationFormValues>()
-}
